utils/pagination.py
from rest_framework.pagination import PageNumberPagination
from rest_framework.response import Response
from functools import wraps
from django.db.models import QuerySet


class CustomPagination(PageNumberPagination):
    page_size_query_param = "page_size"
    page_size = 18

    def get_paginated_response(self, data):
        return Response(
            {
                "next": self.get_next_link(),
                "previous": self.get_previous_link(),
                "count": self.page.paginator.count,
                "results": data,
            }
        )


def paginate(func):
    @wraps(func)
    def inner(self, *args, **kwargs):
        queryset = func(self, *args, **kwargs)
        assert isinstance(
            queryset, (list, QuerySet)
        ), "apply_pagination expects a List or a QuerySet"

        page = self.paginate_queryset(queryset)
        if page is not None:
            serializer = self.get_serializer(page, many=True)
            return self.get_paginated_response(serializer.data)

        serializer = self.get_serializer(queryset, many=True)
        return Response(serializer.data)

    return inner
